+++ eta-company-calendar-frontend/src/app/user-management/user-management.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { TranslateModule } from '@ngx-translate/core';
 import { SharedModule } from '~/app/shared/shared.module';
 import { UserManagementRoutingModule } from '../user-management/user-management-routing.module';
-import { UserListComponent } from './components/user-list.component';
 import { UserManagementDescriptionComponent } from './components/user-management-description.component';
 import { DeleteUserComponent } from './modals/delete-user.component';
 import { EditUserComponent } from './modals/edit-user.component';
@@ -14,7 +13,6 @@ import { UserService } from './service/user-service';
     UserManagementDescriptionComponent,
     NewUserComponent,
     EditUserComponent,
-    UserListComponent,
     DeleteUserComponent,
   ],
   imports: [
